Move homeModule
