+++ ui/home/src/main/kotlin/com/github/kr328/clash/home/di/HomeModule.kt
@@ -1,4 +1,4 @@
-package com.github.kr328.clash.home
+package com.github.kr328.clash.home.di
 
 import com.github.kr328.clash.home.vm.AndroidDependencies
 import com.github.kr328.clash.home.vm.HomeViewModel
